requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
python-jose
pydantic
psycopg
pydantic-settings
pydantic[email]
bcrypt==4.3.0
passlib[bcrypt]==1.7.4
pytest
pytest-asyncio
pytest-cov
httpx
python-multipart
black
